Modal contents now are new components, functions to manage their respective insertions will be done on the component
